+++ iommu_ref_model/test/test_app.c
@@ -164,7 +164,7 @@ main(void) {
     exp_msg.EXEC_REQ = 0;
     exp_msg.DSV = 1;
     exp_msg.DSEG = 0x43;
-    exp_msg.PAYLOAD = (0x1234UL << 48UL) | (PRGR_RESPONSE_FAILURE << 44UL);
+    exp_msg.PAYLOAD = (0x1234UL << 48UL) | ((uint64_t)PRGR_RESPONSE_FAILURE << 44UL);
     handle_page_request(&iommu, &pr);
     fail_if( ( exp_msg_received == 0 ) );
     fail_if( ( check_faults(&iommu, 256, pr.PV, pr.PID, pr.PRIV, 0x431234, PAGE_REQ_MSG_CODE, PCIE_MESSAGE_REQUEST, 0) < 0 ) );
@@ -205,7 +205,7 @@ main(void) {
     exp_msg.EXEC_REQ = 0;
     exp_msg.DSV = 1;
     exp_msg.DSEG = 0x43;
-    exp_msg.PAYLOAD = (0x1234UL << 48UL) | (PRGR_INVALID_REQUEST << 44UL);
+    exp_msg.PAYLOAD = (0x1234UL << 48UL) | ((uint64_t)PRGR_INVALID_REQUEST << 44UL);
     handle_page_request(&iommu, &pr);
     fail_if( ( exp_msg_received == 0 ) );
     fail_if( ( check_faults(&iommu, 260, pr.PV, pr.PID, pr.PRIV, 0x431234, PAGE_REQ_MSG_CODE, PCIE_MESSAGE_REQUEST, 0) < 0 ) );
@@ -237,7 +237,7 @@ main(void) {
     exp_msg.EXEC_REQ = 0;
     exp_msg.DSV = 1;
     exp_msg.DSEG = 0x43;
-    exp_msg.PAYLOAD = (0x1234UL << 48UL) | (PRGR_INVALID_REQUEST << 44UL);
+    exp_msg.PAYLOAD = (0x1234UL << 48UL) | ((uint64_t)PRGR_INVALID_REQUEST << 44UL);
     handle_page_request(&iommu, &pr);
     fail_if( ( exp_msg_received == 0 ) );
     fail_if( ( check_faults(&iommu, 260, pr.PV, pr.PID, pr.PRIV, 0x431234, PAGE_REQ_MSG_CODE, PCIE_MESSAGE_REQUEST, 0) < 0 ) );
@@ -268,7 +268,7 @@ main(void) {
     exp_msg.EXEC_REQ = 0;
     exp_msg.DSV = 1;
     exp_msg.DSEG = 0x43;
-    exp_msg.PAYLOAD = (0x1234UL << 48UL) | (PRGR_INVALID_REQUEST << 44UL);
+    exp_msg.PAYLOAD = (0x1234UL << 48UL) | ((uint64_t)PRGR_INVALID_REQUEST << 44UL);
     handle_page_request(&iommu, &pr);
     fail_if( ( exp_msg_received == 0 ) );
     fail_if( ( check_faults(&iommu, 260, pr.PV, pr.PID, pr.PRIV, 0x431234, PAGE_REQ_MSG_CODE, PCIE_MESSAGE_REQUEST, 0) < 0 ) );
@@ -294,7 +294,7 @@ main(void) {
     exp_msg.EXEC_REQ = 0;
     exp_msg.DSV = 1;
     exp_msg.DSEG = 0x43;
-    exp_msg.PAYLOAD = (0x1234UL << 48UL) | (PRGR_INVALID_REQUEST << 44UL);
+    exp_msg.PAYLOAD = (0x1234UL << 48UL) | ((uint64_t)PRGR_INVALID_REQUEST << 44UL);
     handle_page_request(&iommu, &pr);
     fail_if( ( exp_msg_received == 0 ) );
     fail_if( ( check_faults(&iommu, 260, pr.PV, pr.PID, pr.PRIV, 0x431234, PAGE_REQ_MSG_CODE, PCIE_MESSAGE_REQUEST, 0) < 0 ) );
@@ -3125,7 +3125,7 @@ main(void) {
     exp_msg.EXEC_REQ = 0;
     exp_msg.DSV = 1;
     exp_msg.DSEG = 0x43;
-    exp_msg.PAYLOAD = (0x1234UL << 48UL) | (PRGR_RESPONSE_FAILURE << 44UL);
+    exp_msg.PAYLOAD = (0x1234UL << 48UL) | ((uint64_t)PRGR_RESPONSE_FAILURE << 44UL);
     handle_page_request(&iommu, &pr);
     fail_if( ( exp_msg_received == 0 ) );
     fail_if( ( check_faults(&iommu, 258, pr.PV, pr.PID, pr.PRIV, 0x431234, PAGE_REQ_MSG_CODE, PCIE_MESSAGE_REQUEST, 0) < 0 ) );
@@ -3141,7 +3141,7 @@ main(void) {
     exp_msg.DSEG = 0x11;
     exp_msg.PV = 0;
     exp_msg.PID = 0;
-    exp_msg.PAYLOAD = (0x2233UL << 48UL) | (PRGR_INVALID_REQUEST << 44UL);
+    exp_msg.PAYLOAD = (0x2233UL << 48UL) | ((uint64_t)PRGR_INVALID_REQUEST << 44UL);
     handle_page_request(&iommu, &pr);
     fail_if( ( exp_msg_received == 0 ) );
     fail_if( ( check_faults(&iommu, 260, pr.PV, pr.PID, pr.PRIV, 0x112233, PAGE_REQ_MSG_CODE, PCIE_MESSAGE_REQUEST, 0) < 0 ) );
@@ -3157,7 +3157,7 @@ main(void) {
     exp_msg.DSEG = 0x11;
     exp_msg.PV = 0;
     exp_msg.PID = 0;
-    exp_msg.PAYLOAD = (0x2233UL << 48UL) | (PRGR_INVALID_REQUEST << 44UL);
+    exp_msg.PAYLOAD = (0x2233UL << 48UL) | ((uint64_t)PRGR_INVALID_REQUEST << 44UL);
     handle_page_request(&iommu, &pr);
     fail_if( ( exp_msg_received == 0 ) );
     fail_if( ( check_faults(&iommu, 260, pr.PV, pr.PID, pr.PRIV, 0x112233, PAGE_REQ_MSG_CODE, PCIE_MESSAGE_REQUEST, 0) < 0 ) );
@@ -3174,7 +3174,7 @@ main(void) {
     exp_msg.DSEG = 0x11;
     exp_msg.PV = 1;
     exp_msg.PID = 0xBABEC;
-    exp_msg.PAYLOAD = (0x2233UL << 48UL) | (PRGR_RESPONSE_FAILURE << 44UL);
+    exp_msg.PAYLOAD = (0x2233UL << 48UL) | ((uint64_t)PRGR_RESPONSE_FAILURE << 44UL);
     handle_page_request(&iommu, &pr);
     fail_if( ( exp_msg_received == 0 ) );
     fail_if( ( ((read_register(&iommu, PQH_OFFSET, 4)) != read_register(&iommu, PQT_OFFSET, 4)) ) );
@@ -3217,7 +3217,7 @@ main(void) {
     exp_msg.DSEG = 0x11;
     exp_msg.PV = 0;
     exp_msg.PID = 0;
-    exp_msg.PAYLOAD = (0x2233UL << 48UL) | (PRGR_SUCCESS << 44UL);
+    exp_msg.PAYLOAD = (0x2233UL << 48UL) | ((uint64_t)PRGR_SUCCESS << 44UL);
     handle_page_request(&iommu, &pr);
     fail_if( ( message_received == 0 ) );
     fail_if( ( exp_msg_received == 0 ) );
@@ -3252,7 +3252,7 @@ main(void) {
     access_viol_addr = (pqb.ppn * PAGESIZE);
     access_viol_addr += (read_register(&iommu, PQT_OFFSET, 4) * 16);
     pr.PAYLOAD = 0xdeadbeef00000007; // Set last = 1, PRG index = 0
-    exp_msg.PAYLOAD = (0x2233UL << 48UL) | (PRGR_RESPONSE_FAILURE << 44UL);
+    exp_msg.PAYLOAD = (0x2233UL << 48UL) | ((uint64_t)PRGR_RESPONSE_FAILURE << 44UL);
     message_received = 0;
     handle_page_request(&iommu, &pr);
     fail_if( ( message_received == 0 ) );
